LayoutTests/http/tests/appcache/interrupted-update-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 37: Reloading...
CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
Test that an application cache update works again after interruption.

Reloaded
CHECKING
DOWNLOADING
PROGRESS 0/9
PROGRESS 1/9
PROGRESS 2/9
PROGRESS 3/9
PROGRESS 4/9
PROGRESS 5/9
PROGRESS 6/9
PROGRESS 7/9
PROGRESS 8/9
PROGRESS 9/9
CACHED

PASS
